// src/generated/prisma/enums.ts
export const SocialPlatform = {
  FACEBOOK: 'FACEBOOK',
  LINKEDIN: 'LINKEDIN'
} as const

export type SocialPlatform = (typeof SocialPlatform)[keyof typeof SocialPlatform]
// ...
export const PostTargetStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  OVERDUE: 'OVERDUE',
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  PUBLISHED: 'PUBLISHED',
  BLOCKED: 'BLOCKED'
} as const

export type PostTargetStatus = (typeof PostTargetStatus)[keyof typeof PostTargetStatus]